+++ db/migrate/016_add_is_closed_to_purchase_request_statuses.rb
@@ -1,0 +1,13 @@
+class AddIsClosedToPurchaseRequestStatuses < ActiveRecord::Migration[6.0]
+  def up
+    unless column_exists?(:purchase_request_statuses, :is_closed)
+      add_column :purchase_request_statuses, :is_closed, :boolean, default: false, null: false
+    end
+  end
+  
+  def down
+    if column_exists?(:purchase_request_statuses, :is_closed)
+      remove_column :purchase_request_statuses, :is_closed
+    end
+  end
+end